Use abc.Mapping for improved typing, close #2
This fixes mypy issues with `list(name)` and `dict(name)` where
we partially implemented the mapping protocol but didn't use the
actual abstract base class. Now we do. The cost is that `list(name)`
must now be `name.values()` but that's more consistent with other
mappings and `list(name)` is not the primary use case.