Remove single-string output as a design goal
Previous documentation and testing aimed for idempotent string
output such that str(Name(x)) == str(Name(str(Name(x)))). Upon
further consideration, this is not a useful design goal for a name
parser which is focused on the separation of name fields, not the
recombination. This led to corner cases -- such as dividing first
and middle name fields with the presence of a comma -- which would
promote less intuitive string output, calling the question of what
purpose the single-string output has.

This update removes idempotence and all priority on single strings.
Single strings may still be useful at a glance, and are still part
of the package going forward, but will not be prioritized in any
case over the best possible dict parsing.