Create skip_cleaning and prefixes Name arguments
Both ideas are discussed in #58.

- skip_cleaning completely skips multiple stages of cleaning names
and name parts.
- prefixes allows the user to modify the list of prefixes, exposed
from the default config.PREFIXES

This currently only affects the Name class and is not part of the
CLI or a full release while I consider:

1. How to implement a general way to override config lists such as
suffixes, conjunctions, titles, etc. It's currently possible only
to monkeypatch these, but since we can now adjust prefixes we
should extend this to other constants.
2. Whether and how to incorporate any of these changes into the
CLI. I'm leaning toward no, but perhaps --skip-cleaning is fine.